VisualSearchBar: update entity previews ondropdownleave
adds onDropdownLeave handler to InputDropdown, so that
VisualSearchBar can update entity previews when the navigates
back up from an autocomplete suggestion back to their original query.

This could not go into onInputChange, because that would result in dupicate
calls to updateEntityPreviews when the user types. This wouldn't be an issue right now,
due to our debouncing, but could cause future problems.

J=SLAP-1747
TEST=manual

test that entity previews update when navigating back up
test that updateEntityPreviews is not called multiple times when the user types
or navigates with the arrow keys