configurable options per facet (#13)
To match the customization level in SDK, this pr update Facets and Facet component to allow configuration of individual facet for following options: searchable, collapsible, defaultExpanded, label, and placeholderText. If not provided, Facet component will use the top-level configuration for Facets and any default value in place.

J=SLAP-1764
TEST=manual

add facetConfigs to Facets in LocationPage with different configuration combo, and see that it works as expected.