Add FilterDisplayManager (#30)
Add `FilterDisplayManager` for showing and hiding the filters based on the screen width and the view/hide filter buttons

A `PageView` is added to context through the `PageViewContextProvider`, and any component or page can access or modify it. There is a `FilterDisplayManager` which renders the filters based on the page view. Filters and dividers are passed to it using React props children. On desktop, it displays the filters, and on mobile it collapses them. If the `ViewFiltersButton` is pressed, the filters will expand to fill the whole page.

The mocks show that the filter dividers in the mobile expanded filter view goes to the edge of the screen. In order to accomplish this, I modify the styling of any `Facets`, `StaticFilters`, or `Divider` components inside the `FilterDisplayManager` so that the divider takes up the full width. Unfortunately this means that those components inside the `CollapsibleFiltersDrawer` are forced to use the 'assign' CompositionMethod. It is worth considering alternative approaches in the future on how we can override styling out-of-the-box and also allow the user to provide any kind of css `CompositionMethod`. Also, the Applied Filters need different styling when they are used in the universal results component. This is another use case where the ability to apply custom styling within the search starter and cleanly merge it with the user defined styling would be helpful. For now, I opted to copy the styling into the `SectionHeader` and make it customizable that way. This does introduce a little but of duplication to the styling, however I think that is preferable to modifying to the component on a vertical and having unintended conflicts on the universal page.

J=SLAP-1774
TEST=manual

Test on both mobile and and desktop and various screen sizes. See that on desktop the filters are visible. On mobile, the filters hide. I can press the button to see the filters, and scroll if needed. I can click the close button to go back to the mobile view. Resizing the window from the FiltersVisibleMobile to Desktop will properly update to the Desktop view.