Consolidate SearchBar and VisualSearchBar using Dropdown (#52)
This commit consolidates the two SearchBar's we have, by using the newly added Dropdown component.

As part of this, Dropdown was updated to no longer need a `numItems` prop on the main component,
and `index` props for `DropdownItem`s.
This was done by auto-magically recursing through child components.
I don't believe this is an inherently bad pattern, it just needs to be applied carefully.
One difference between this pattern's usage here, vs how it was used for the original InputDropdown, is that we don't expose the injected props to the user. Instead, there is an external facing component, DropdownItem, and an internal DropdownItemWithIndex that is used internally by Dropdown.

This change was necessary due to the complexity of the DropdownItems necessary for VisualAutocomplete.
Making user's manually specify them is fine for simple cases, and is almost certainly more performant, but for more complex cases with multiple arrays of data, it quickly becomes hard to juggle between which item is which index, and oddly restrictive in terms code formatting.
I did not notice any performance regressions from recursing through props.children.

In order to support verticalLinks, an arbitrary "metadata" attribute was added to DropdownItem.
I initially attempted to use generics for this, but found them unwieldy and not terribly helpful.
I don't foresee us needing to store massive amounts of metadata.

I also removed the ability to specify your own headlessId. It doesn't seem particularly useful, and clutters up the props, especially since it's completely unused for the regular SearchBar case.

Removed the `VisualAutocomplete/` folder, moved `EntityPreviews.tsx` and `SampleVisualSearchBar.tsx` down a level.
`filteredRecentSearches` is no longer its own state variable because it's not necessary.

J=SLAP-1777
TEST=manual

tested vertical links, both by clicking on them with the mouse, and focusing them with the arrow keys and hitting enter (tested on the slanswers experience since the answers-js-docs does not have sufficient test data)

vertical links do not appear for vertical search pages

tested selecting recent searches, and that they have the correct styling when focused with the arrow keys
recent searches do not show up on vertical search pages
recent searches are still filtered by the current query and levenshtein distance