Serve on Yext Sites (#62)
Serve on yext sites

This adds default configuration to support yext sites. It works for both full production builds and for live preview. Currently the page is being served by the dev branch, but it will be available for develop and main once it's merged into those branches. I opted for `npm install` for the `installDepsCmd` so that the package.json file can be modified in live preview and the package-lock.json can be updated as necessary. For the `buildSetupCmd`, I went with `npm ci` because it's faster and ensures we get a clean build based purely on the package-lock.json file.

J=SLAP-1841
TEST=manual

Add the repo to yext sites and view the page being served at https://devyextsites-react--site--search--starter-slapshot-pagescdn-com.preview.pagescdn.com/. Also load the experience in the code editor and see the page work with live preview. Modify a file and see the file reflected in the page after refreshing the page