Support SCSS customization (#61)
Adds support for customizing the app's styling by changing the common styles in the PB config. The style variables are computed in the `PageBuilderApp` any time the config changes. Refreshing the page on the app displays the updated styling.

For now, the styling only affects the `LocationBias`, `StandardCard`, and `SpellCheck` components. The styling for these has been translated from tailwind to SCSS because PB 2.0 will not be using tailwind.

J=SLAP-1788
TEST=manual

Check that updating a style in the config is reflected when refreshing the page.