fix navigation styling from tailwind 2->3 upgrade
Upgrading tailwind from 2 to 3 introduced a number of styling regressions
due to a number of arguably breaking but undocumented changes.

It looks like tailwind has changed how it behaves when two tailwind classes modify
the same css attribute, for example having text-blue-600 and text-gray-800 on the same element.
The Navigation component was updated to no longer rely on this kind of behavior.

Tailwind also seemingly no longer supports border-opacity - it is no longer present
in the v3 docs, despite not being mentioned as removed in the upgrade guide.

J=SLAP-1864
TEST=manual

test side by side with local site and a version of the react-site-search-starter
hosted on yext sites that doesn't have this styling regression

the navigation's menu button (which appears when the navigation tabs no longer fully
fit within the page's width) has correct styling for it's 4 different states

navigation tab styling outside of the menu have correct styling for both active, inactive, and hover
states