Fix VA styling (#67)
Fix a styling regression where the magnifying glass icon for query suggestions in VA no longer appeared.

Also, fix the styling for entity previews when no other items are in the dropdown. In this case, the divider should not appear for the entity previews of the first vertical. But, if there are other verticals with entity previews, then these should have dividers.

J=SLAP-1861
TEST=manual

See that the icons for VA are correct and the divider above entity previews only appears over a vertical's preview cards if there are other dropdown items before it.