Refactor StaticFilters and Facets (#68)
Refactor StaticFilters and Facets to use a common Filters component

The common Filters component allows all of the styling and functionality (collapsible and searchable) to be reused. The purpose of the StaticFilters and Facets components is now to gather the filters data and click handlers and pass them to the Filters component. The Filters components renders the possible filters for various fields, and the CollapsibleLabel allows the Filters to be collapsed.

The facets have a number to indicate the count (e.g. 'Cost (3)'), and this number with the parenthesis will now be used in the Levenshtein matching since only the label is passed to the Filters component. This seems to work just fine with testing.

J=SLAP-1856
TEST=manual

Compare side by side with the hosted starter app and confirm styling is the same. Check desktop and mobile view. Test the custom config for the StaticFilters and Facets and confirm the functionality still works