sync out of date package-lock
The package-lock.json was out of sync.
When the package-lock is not in sync, npm ci cannot be run.
When npm ci cannot be run, our ci does not work.

see: https://github.com/yext/react-site-search-starter/pull/77

This PR runs npm install again. I didn't see any purpose of updating
the package.json version - this repo is not published to npm or anything
like that.

J=None
TEST=None